[WFLY-490] / [WFLY-1864] Update the protocol so that servers receiving management requests from a trusted master can request a Subject to use when executing the operation.
For host controller to managed server interactions the two are always kept in sync version wise.
For master to slave communication the master must be the same version as the slave or higher i.e. a slave sending this new request will not cause an interop issue as the master must be at least a version that understands it.