JMX/Infinispan/JGroups Capabilities and Requirements
This makes the JMX requirement optional. The JMX requirement is used to support statistical mbeans from the caches and the JGroups channel.

a) The current cache manager dependency on JMX isn't optional, although the code the uses the injection supports having nothing injected.
b) The current JGroups channel dependency on JMX isn't optional. But I changed ChannelConnectorBuilder to not require JMX.